MacAssistant/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0
openai==0.27.0
python-dotenv==1.0.0
requests==2.28.2
flask-socketio==5.3.2
pyyaml==6.0
pytest==7.3.1